Fix some printf formats
* libdb/db_lookup.c (list_extensions): Format the return value of
gl_list_size using %zu, not %zd.
* src/catman.c (add_arg): Format the return values of MYDBM_DSIZE and
strlen using %zu, not %zd.
(parse_for_sec): Format ARG_MAX - arg_size using %zu, not %zd.
* src/man.c (main): Format ruid and euid using %lu, not %d (cast to
unsigned long since POSIX only says that uid_t is an integer type).
* src/manp.c (describe_flag): Format flag using %u, not %d.

+++ src/man.c
@@ -4082,7 +4082,8 @@ int main (int argc, char *argv[])
 
 	get_term (); /* stores terminal settings */
 #ifdef MAN_OWNER
-	debug ("real user = %d; effective user = %d\n", ruid, euid);
+	debug ("real user = %lu; effective user = %lu\n",
+	       (unsigned long) ruid, (unsigned long) euid);
 #endif /* MAN_OWNER */
 
 	/* close this locale and reinitialise if a new locale was
